Fix bug with ignoring verbosity level in logger.
Because of changes in the API synchronous request to the service, the
verbosity level of logging can be compared not with the actual value of the debug
level, but with the address of the generator in the memory, which led to
the installation of the wrong verbosity level of logging.
Now real verbosity level is compared.