Services are able to send more than 640Kb in synchronous calls.
The problem was incomplete send a message to the socket, due to the fact
that the recorded number of bytes in the socket buffer is less data to
be written. Actually the message is not delivered in full, and the
remaining part is not sent.
With asynchronous calls the fullness of sending the message to a socket was controlled by the cycle.
To ensure a full send with the synchronous calls added method write_all for a
synchronous write to a socket that uses socket.sendall().