Fixed an error that occurred while processing multiple sessions of the same events.
The error was related to the fact that objects containing context, are
not copied, but were common. This led to the fact that when the
asynchronous operation of objects may spoil the context of others. Now
the object that holds the context of the session, is created each time
it is started and unique.