[Bug] Fix recursion bug when CocaineHandler is used in root logger.
The framework has a loggin system for debug, this system uses
standart logging module. The logger logs messages about internal
processes of services. Also cocaine Logger service is based on
Service code too. If CocaineHandler is attached to root logger
all mesages from "cocaine.*" loggers pass to this handler. It leads
to call Service code inside the handler and so on. There we meet
a recursion call. To fix this problem propagate in "cocaine" logger
is disabled, so it safe now to use CocaineHandler. Also a test
was adopted for this case.